Q&D fix for flatpak-builder 1.4.3
flathub build server update to flatpak-builder 1.4.3.
A new mecanism introduced in this version of flatpak-builder cause a
mismatch between OpenCV cmake config and the package build tree.

In this commit, we force implicit inclusion of opencv headers directory
to make the build pass and release the 2.13.2 vesion of CC but this
requires further investigation to make a better fix.